Include ristretto255 into the base bundle.
