+++ package.json
@@ -1,15 +1,18 @@
 {
   "name": "noble-ed25519",
   "version": "0.2.2",
-  "description": "Noble ed25519. High-security, easily auditable, 0-dep, 1-file pubkey, scalarmult & EDDSA.",
+  "description": "Noble ed25519 & ristretto255. High-security, easily auditable, 0-dep, 1-file pubkey, scalarmult & EDDSA.",
   "main": "index.js",
   "files": [
     "index.js",
-    "index.d.ts"
+    "index.d.ts",
+    "ristretto255.js",
+    "ristretto255.d.ts"
   ],
   "scripts": {
     "compile": "tsc -d",
-    "test": "jest"
+    "test": "jest",
+    "bench": "node benchmark.js"
   },
   "jest": {
     "testRegex": "/.*?\\.test.ts",
@@ -25,16 +28,19 @@
   },
   "license": "MIT",
   "devDependencies": {
-    "@types/jest": "^24.0.15",
+    "@types/jest": "^24.0.13",
     "@types/node": "^12.0.8",
-    "fast-check": "^1.15.1",
+    "benchmark": "^2.1.4",
+    "fast-check": "^1.16.0",
     "jest": "^24.8.0",
     "ts-jest": "^24.0.2",
-    "typescript": "^3.5.2"
+    "typescript": "^3.4.5"
   },
   "keywords": [
     "ed25519",
     "edwards",
+    "ristretto255",
+    "decaf",
     "curve25519",
     "x25519",
     "elliptic",
