--- ristretto255/jest.config.js
@@ -1,6 +1,0 @@
-module.exports = {
-  testRegex: "/.*?\.test.ts",
-  transform: {
-    "^.+\\.ts$": "ts-jest"
-  }
-};
