--- ristretto255/package.json
@@ -1,42 +1,0 @@
-{
-  "name": "noble-ristretto255",
-  "version": "0.1.0",
-  "description": "Noble ristretto255",
-  "main": "lib/index.js",
-  "files": [
-    "lib"
-  ],
-  "scripts": {
-    "test": "jest",
-    "bench": "node benchmark.js"
-  },
-  "author": "Paul Miller (https://paulmillr.com)",
-  "license": "MIT",
-  "devDependencies": {
-    "@types/jest": "^24.0.13",
-    "@types/node": "^12.0.8",
-    "benchmark": "^2.1.4",
-    "fast-check": "^1.16.0",
-    "jest": "^24.8.0",
-    "ts-jest": "^24.0.2",
-    "typescript": "^3.4.5"
-  },
-  "keywords": [
-    "ristretto255",
-    "decaf",
-    "cofactor elimination",
-    "ed25519",
-    "edwards",
-    "curve25519",
-    "x25519",
-    "elliptic",
-    "elliptic curve",
-    "curve",
-    "signature",
-    "ecc",
-    "eddsa",
-    "ecdsa",
-    "cryptography",
-    "security"
-  ]
-}
