--- ristretto255/tsconfig.json
@@ -1,14 +1,0 @@
-{
-  "compilerOptions": {
-    "module": "commonjs",
-		"strict": true,
-    "target": "esnext",
-    "noImplicitAny": true,
-    "removeComments": true,
-    "preserveConstEnums": true,
-    "experimentalDecorators": true,
-    "outDir": "lib"
-  },
-  "include": ["**/*"],
-  "exclude": ["node_modules", "**/*.test.ts"]
-}
